animation

+++ src/micapolos/tata8/Random.java
@@ -3,11 +3,11 @@ package micapolos.tata8;
 public class Random {
   private Random() {}
 
-  public static float between(float min, float max) {
-    return Math.round(min + Math.random() * (max - min));
+  public static int between(float min, float max) {
+    return (int) Math.round(min + Math.random() * (max - min));
   }
 
-  public static float until(float limit) {
+  public static int until(float limit) {
     return between(0, limit - 1);
   }
 }
